moved evaluate preferable_allowable_vaccine logic and evaluate age to individual modules
