+++ app/models/target_dose.rb
@@ -25,15 +25,7 @@ class TargetDose
     end
   end
 
-  def set_default_values(return_hash, default_hash={})
-    default_hash.each do |default_value_key, default_value|
-      current_value = return_hash[default_value_key]
-      if current_value.nil? || current_value == ''
-        return_hash[default_value_key] = default_value
-      end
-    end
-    return_hash
-  end
+
 
     # Date Administered
     # Patient Immunization History Administered - Dose Count
@@ -47,25 +39,6 @@ class TargetDose
     # when 'Interval'
     # when 'Vaccine Count by Age'
     # when 'Vaccine Count by Date'
-
-  def create_age_date_attributes(evaluation_antigen_series_dose, dob)
-    age_attrs = {}
-    default_values = {
-      max_age_date: '12/31/2999'.to_date,
-      min_age_date: '01/01/1900'.to_date,
-      absolute_min_age_date: '01/01/1900'.to_date
-    }
-    [
-      'absolute_min_age', 'min_age', 'earliest_recommended_age',
-      'latest_recommended_age', 'max_age'
-    ].each do |action|
-      date_action  = action + '_date'
-      age_string   = evaluation_antigen_series_dose.read_attribute(action)
-      patient_date = create_patient_age_date(age_string, dob)
-      age_attrs[date_action.to_sym] = patient_date
-    end
-    set_default_values(age_attrs, default_values)
-  end
 
   def create_interval_date_attributes(interval_object, previous_dose_date)
     interval_attrs = {}
@@ -85,77 +58,12 @@ class TargetDose
     set_default_values(interval_attrs, default_values)
   end
 
-  def create_vaccine_attributes(evaluation_antigen_series_dose_vaccine, dob)
-    vaccine_attrs = {}
-    default_values = {
-      begin_age_date: '01/01/1900'.to_date,
-      end_age_date: '12/31/2999'.to_date
-    }
-    %w(begin_age end_age).each do |action|
-      date_action  = action + '_date'
-      age_string   =
-        evaluation_antigen_series_dose_vaccine.read_attribute(action)
-      patient_date = create_patient_age_date(age_string, dob)
-      vaccine_attrs[date_action.to_sym] = patient_date
-    end
-    vaccine_attrs[:expected_trade_name] =
-      evaluation_antigen_series_dose_vaccine.trade_name
-    vaccine_attrs[:expected_volume] =
-      evaluation_antigen_series_dose_vaccine.volume
-
-    set_default_values(vaccine_attrs, default_values)
-  end
-
   def create_gender_attributes(evaluation_antigen_series_dose)
     gender_attrs = {}
     default_values = {}
     gender_attrs[:required_gender] =
       evaluation_antigen_series_dose.required_gender
     set_default_values(gender_attrs, default_values)
-  end
-
-  def get_age_status(age_evaluation_hash,
-                     antigen_administered_record,
-                     previous_dose_status_hash=nil)
-    # As described on page 38 (TABLE 4 - 12) in the CDC logic specifications
-    age_status = {record: antigen_administered_record}
-    if age_evaluation_hash[:absolute_min_age] == false
-      age_status[:status]  = 'invalid'
-      age_status[:reason]  = 'age'
-      age_status[:details] = 'too_young'
-    elsif age_evaluation_hash[:min_age] == false
-      has_previous_dose = !previous_dose_status_hash.nil?
-      is_valid = true
-
-      if has_previous_dose
-        previous_dose_invalid = previous_dose_status_hash[:status] == 'invalid'
-        previous_dose_reason  = previous_dose_status_hash[:reason]
-        age_or_interval = ['age', 'interval'].include?(previous_dose_reason)
-
-        if previous_dose_invalid && age_or_interval
-          is_valid = false
-        end
-      end
-
-      if is_valid
-        age_status[:status]  = 'valid'
-        age_status[:reason]  = 'grace_period'
-      else
-        age_status[:status]  = 'invalid'
-        age_status[:reason]  = 'age'
-        age_status[:details] = 'too_young'
-      end
-
-    elsif age_evaluation_hash[:max_age] == false
-      # Should we include extraneous on this as well? Where?
-      age_status[:status]  = 'invalid'
-      age_status[:reason]  = 'age'
-      age_status[:details] = 'too_old'
-    else
-      age_status[:status]  = 'valid'
-      age_status[:reason]  = 'on_schedule'
-    end
-    age_status
   end
 
   def get_interval_status(interval_evaluation_hash,
@@ -194,44 +102,6 @@ class TargetDose
     interval_status
   end
 
-  def get_preferable_vaccine_status(vaccine_evaluation_hash,
-                                        previous_dose_status_hash=nil)
-    vaccine_status = {}
-    if vaccine_evaluation_hash[:begin_age] == false ||
-       vaccine_evaluation_hash[:end_age] == false
-        vaccine_status[:status] = 'invalid'
-        vaccine_status[:reason] = 'preferable'
-        vaccine_status[:details] = 'out_of_age_range'
-    elsif vaccine_evaluation_hash[:trade_name] == false
-      vaccine_status[:status] = 'invalid'
-      vaccine_status[:reason] = 'preferable'
-      vaccine_status[:details] = 'wrong_trade_name'
-    elsif vaccine_evaluation_hash[:volume] == false
-      vaccine_status[:status] = 'valid'
-      vaccine_status[:reason] = 'preferable'
-      vaccine_status[:details] = 'less_than_recommended_volume'
-    else
-      vaccine_status[:status] = 'valid'
-      vaccine_status[:reason] = 'preferable'
-    end
-    vaccine_status
-  end
-
-  def get_allowable_vaccine_status(vaccine_evaluation_hash,
-                                   previous_dose_status_hash=nil)
-    vaccine_status = {}
-    if vaccine_evaluation_hash[:begin_age] == false ||
-       vaccine_evaluation_hash[:end_age] == false
-        vaccine_status[:status] = 'invalid'
-        vaccine_status[:reason] = 'allowable'
-        vaccine_status[:details] = 'out_of_age_range'
-    else
-      vaccine_status[:status] = 'valid'
-      vaccine_status[:reason] = 'allowable'
-    end
-    vaccine_status
-  end
-
   def get_gender_status(gender_evaluation_hash)
     gender_status = {}
     gender_status[:reason] = 'gender'
@@ -265,34 +135,6 @@ class TargetDose
 # dose administered which has evaluation status “Valid” or “Not
 # Valid” if from immediate previous dose administered is “Y”.
 
-
-  def evaluate_dose_age(age_date_attrs, date_of_dose)
-    evaluated_hash = {}
-    [
-      'absolute_min_age_date',
-      'min_age_date',
-      'earliest_recommended_age_date',
-      'latest_recommended_age_date',
-      'max_age_date'
-    ].each do |age_attr|
-      result = nil
-      if !age_date_attrs[age_attr.to_sym].nil?
-        if ['latest_recommended_age_date', 'max_age_date'].include?(age_attr)
-          result = validate_date_equal_or_before(
-                     age_date_attrs[age_attr.to_sym],
-                     date_of_dose
-                   )
-        else
-          result = validate_date_equal_or_after(age_date_attrs[age_attr.to_sym],
-                                                date_of_dose)
-        end
-      end
-      result_attr = age_attr.split('_')[0..-2].join('_')
-      evaluated_hash[result_attr.to_sym] = result
-    end
-    evaluated_hash
-  end
-
   def evaluate_interval_dates(interval_date_attrs, date_of_second_dose)
     evaluated_hash = {}
     %w(
@@ -318,38 +160,6 @@ class TargetDose
       result_attr = interval_attr.split('_')[0..-2].join('_')
       evaluated_hash[result_attr.to_sym] = result
     end
-    evaluated_hash
-  end
-
-  def evaluate_vaccine_attributes(vaccine_attrs, date_of_dose,
-                                  dose_trade_name, dose_volume='0')
-    dose_volume = '0' if dose_volume == '' || dose_volume.nil?
-    evaluated_hash = {}
-    %w(
-      begin_age_date
-      end_age_date
-    ).each do |vaccine_attr|
-      result = nil
-      if !vaccine_attrs[vaccine_attr.to_sym].nil?
-        if vaccine_attr == 'end_age_date'
-          result = validate_date_equal_or_before(
-                     vaccine_attrs[vaccine_attr.to_sym],
-                     date_of_dose
-                   )
-        else
-          result = validate_date_equal_or_after(
-                     vaccine_attrs[vaccine_attr.to_sym],
-                     date_of_dose
-                   )
-        end
-      end
-      result_attr = vaccine_attr.split('_')[0..-2].join('_')
-      evaluated_hash[result_attr.to_sym] = result
-    end
-    evaluated_hash[:trade_name] =
-      vaccine_attrs[:expected_trade_name] == dose_trade_name
-    evaluated_hash[:volume] =
-      vaccine_attrs[:expected_volume].to_f <= dose_volume.to_f
     evaluated_hash
   end
