+++ spec/models/target_dose_spec.rb
@@ -93,55 +93,6 @@ RSpec.describe TargetDose, type: :model do
       end
     end
 
-    describe '#create_age_date_attributes' do
-      %w(absolute_min_age_date min_age_date earliest_recommended_age_date
-         latest_recommended_age_date max_age_date).each do |age_attribute|
-        it "creates a hash with the attribute #{age_attribute}" do
-          age_attrs = test_target_dose.create_age_date_attributes(
-            as_dose,
-            test_patient.dob
-          )
-          as_dose_attribute  = age_attribute.split('_')[0...-1].join('_')
-          as_dose_age_string = as_dose.send(as_dose_attribute)
-          age_date = test_target_dose.create_patient_age_date(
-            as_dose_age_string, test_patient.dob
-          )
-          expect(age_attrs[age_attribute.to_sym]).to eq(age_date)
-          expect(age_attrs[age_attribute.to_sym].class.name).to eq('Date')
-        end
-      end
-      describe 'default values' do
-        # As described on page 38 on CDC logic specs
-        # 'http://www.cdc.gov/vaccines/programs/iis/interop-proj/'\
-        #   'downloads/logic-spec-acip-rec.pdf'
-
-        it 'sets default value for max_age_date' do
-          as_dose.max_age = nil
-          age_attrs = test_target_dose.create_age_date_attributes(
-            as_dose,
-            test_patient.dob
-          )
-          expect(age_attrs[:max_age_date]).to eq('12/31/2999'.to_date)
-        end
-        it 'sets default value for min_age_date' do
-          as_dose.min_age = nil
-          age_attrs = test_target_dose.create_age_date_attributes(
-            as_dose,
-            test_patient.dob
-          )
-          expect(age_attrs[:min_age_date]).to eq('01/01/1900'.to_date)
-        end
-        it 'sets default value for absolute_min_age_date' do
-          as_dose.absolute_min_age = nil
-          age_attrs = test_target_dose.create_age_date_attributes(
-            as_dose,
-            test_patient.dob
-          )
-          expect(age_attrs[:absolute_min_age_date]).to eq('01/01/1900'.to_date)
-        end
-      end
-    end
-
     xdescribe '#age_eligible?' do
       it 'sets @eligible? attribute to true if target_dose is eligible' do
         expect(test_target_dose.eligible).to eq(nil)
@@ -225,193 +176,6 @@ RSpec.describe TargetDose, type: :model do
         it 'returns false if there is no conditional skip' do
           expect(target_dose_no_cond_skip.has_conditional_skip?).to be(false)
         end
-      end
-    end
-
-    describe '#evaluate_dose_age' do
-      let(:valid_age_attrs) do
-        {
-          absolute_min_age_date: 12.months.ago.to_date,
-          min_age_date: 10.months.ago.to_date,
-          earliest_recommended_age_date: 8.months.ago.to_date,
-          latest_recommended_age_date: 2.months.ago.to_date,
-          max_age_date: 1.month.ago.to_date
-        }
-      end
-      it 'returns a hash that includes \'_date\' in the keys' do
-        dose_date = 6.months.ago.to_date
-        eval_hash = test_target_dose.evaluate_dose_age(valid_age_attrs,
-                                                       dose_date)
-        expect(eval_hash.class).to eq(Hash)
-        eval_hash.each do |key, _value|
-          expect(key.to_s.include?('_date')).to eq(true)
-        end
-      end
-      describe 'for each minimum age attribute' do
-        dose_date = 9.months.ago.to_date
-        attribute_options = {
-          before_the_dose_date: [10.months.ago.to_date, true],
-          after_the_dose_date: [8.months.ago.to_date, false],
-          nil: [nil, nil]
-        }
-        %w(absolute_min_age_date min_age_date earliest_recommended_age_date)
-          .each do |attribute|
-            attribute_options.each do |descriptor, value|
-              descriptor_string = "returns #{value[1]} when the #{attribute}"\
-                                  " attribute is #{descriptor}"
-              it descriptor_string do
-                valid_age_attrs[attribute.to_sym] = value[0]
-                eval_hash = test_target_dose.evaluate_dose_age(valid_age_attrs,
-                                                               dose_date)
-                expect(eval_hash[attribute.to_sym]).to eq(value[1])
-              end
-            end
-          end
-      end
-      describe 'for each max age attribute' do
-        dose_date = 9.months.ago.to_date
-        attribute_options = {
-          before_the_dose_date: [10.months.ago.to_date, false],
-          after_the_dose_date: [8.months.ago.to_date, true],
-          nil: [nil, nil]
-        }
-        %w(latest_recommended_age_date max_age_date)
-          .each do |attribute|
-            attribute_options.each do |descriptor, value|
-              descriptor_string = "returns #{value[1]} when the #{attribute}"\
-                                  " attribute is #{descriptor}"
-              it descriptor_string do
-                valid_age_attrs[attribute.to_sym] = value[0]
-                eval_hash = test_target_dose.evaluate_dose_age(valid_age_attrs,
-                                                               dose_date)
-                expect(eval_hash[attribute.to_sym]).to eq(value[1])
-              end
-            end
-          end
-      end
-    end
-
-    describe '#get_age_status' do
-      # This logic is defined on page 38 of the CDC logic spec
-      it 'returns invalid, too young for absolute_min_age false' do
-        prev_status_hash = nil
-        age_eval_hash = {
-          absolute_min_age: false,
-          min_age: false,
-          earliest_recommended_age: false,
-          latest_recommended_age: true,
-          max_age: true
-        }
-        expected_result = { status: 'invalid',
-                            reason: 'age',
-                            details: 'too_young',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
-      end
-
-      it 'returns invalid, too young for before min_age and previous invalid' do
-        prev_status_hash = {
-          status: 'invalid',
-          reason: 'age',
-          details: 'too_young'
-        }
-        age_eval_hash = {
-          absolute_min_age: true,
-          min_age: false,
-          earliest_recommended_age: false,
-          latest_recommended_age: true,
-          max_age: true
-        }
-        expected_result = { status: 'invalid',
-                            reason: 'age',
-                            details: 'too_young',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
-      end
-
-      it 'returns valid, grace_period for before min_age and previous valid' do
-        prev_status_hash = {
-          status: 'valid',
-          reason: 'grace_period'
-        }
-        age_eval_hash = {
-          absolute_min_age: true,
-          min_age: false,
-          earliest_recommended_age: false,
-          latest_recommended_age: true,
-          max_age: true
-        }
-        expected_result = { status: 'valid',
-                            reason: 'grace_period',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
-      end
-      it 'returns valid, grace_period for before min_age yet first dose' do
-        prev_status_hash = nil
-        age_eval_hash = {
-          absolute_min_age: true,
-          min_age: false,
-          earliest_recommended_age: false,
-          latest_recommended_age: true,
-          max_age: true
-        }
-        expected_result = { status: 'valid',
-                            reason: 'grace_period',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
-      end
-      it 'returns valid for after min_age and before max_age' do
-        prev_status_hash = nil
-        age_eval_hash = {
-          absolute_min_age: true,
-          min_age: true,
-          earliest_recommended_age: false,
-          latest_recommended_age: true,
-          max_age: true
-        }
-        expected_result = { status: 'valid',
-                            reason: 'on_schedule',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
-      end
-      it 'returns invalid, too_old for after max_age' do
-        prev_status_hash = nil
-        age_eval_hash = {
-          absolute_min_age: true,
-          min_age: true,
-          earliest_recommended_age: true,
-          latest_recommended_age: false,
-          max_age: false
-        }
-        expected_result = { status: 'invalid',
-                            reason: 'age',
-                            details: 'too_old',
-                            record: as_dose }
-        expect(
-          test_target_dose.get_age_status(age_eval_hash,
-                                          as_dose,
-                                          prev_status_hash)
-        ).to eq(expected_result)
       end
     end
 
@@ -714,388 +478,8 @@ RSpec.describe TargetDose, type: :model do
       end
     end
     describe 'implementation of evaluate preferable/allowable vaccines logic' do
-      let(:as_dose_vaccine) { FactoryGirl.create(:antigen_series_dose_vaccine) }
       describe '#evaluate_preferable_vaccine' do
-        # This logic is defined on page 48 of the CDC logic spec to evaluate the
-        # preferable vaccines and if they have been used (or if allowable has
-        # been used)
-      end
 
-      describe '#create_vaccine_attributes' do
-        it 'returns a hash with 4 attributes' do
-          dob = 1.year.ago.to_date
-          as_dose_vaccine.begin_age = '6 weeks'
-          expected_begin_age = dob + 6.weeks
-          as_dose_vaccine.end_age = '5 years'
-          expected_end_age = dob + 5.years
-
-          as_dose_vaccine.trade_name = 'test'
-          expected_trade_name = 'test'
-          as_dose_vaccine.volume = '0.5'
-          expected_volume = '0.5'
-          vaccine_attrs = test_target_dose.create_vaccine_attributes(
-            as_dose_vaccine,
-            dob
-          )
-          expect(vaccine_attrs[:expected_trade_name]).to eq(expected_trade_name)
-          expect(vaccine_attrs[:expected_volume]).to eq(expected_volume)
-          expect(vaccine_attrs[:begin_age_date])
-            .to eq(expected_begin_age)
-          expect(vaccine_attrs[:end_age_date])
-            .to eq(expected_end_age)
-        end
-        %w(
-          begin_age_date end_age_date
-        ).each do |vaccine_attr|
-          it "creates a hash with the attribute #{vaccine_attr}" do
-            dob = 2.years.ago.to_date
-            vaccine_attrs = test_target_dose.create_vaccine_attributes(
-              as_dose_vaccine,
-              dob
-            )
-            abr_attribute = vaccine_attr.split('_')[0...-1].join('_')
-            vaccine_time_string = as_dose_vaccine.send(abr_attribute)
-            expect(vaccine_time_string.nil?).to be(false)
-            vaccine_age_date = test_target_dose.create_patient_age_date(
-              vaccine_time_string, dob
-            )
-            expect(vaccine_attrs[vaccine_attr.to_sym])
-              .to eq(vaccine_age_date)
-            expect(vaccine_attrs[vaccine_attr.to_sym].class.name)
-              .to eq('Date')
-          end
-        end
-        describe 'default values' do
-          # As described on page 38 on CDC logic specs
-          # 'http://www.cdc.gov/vaccines/programs/iis/interop-proj/'\
-          #   'downloads/logic-spec-acip-rec.pdf'
-          it 'sets default value for begin_age_date' do
-            dob = 2.years.ago.to_date
-            as_dose_vaccine.begin_age = nil
-            vaccine_attrs = test_target_dose.create_vaccine_attributes(
-              as_dose_vaccine,
-              dob
-            )
-            expect(
-              vaccine_attrs[:begin_age_date]
-            ).to eq('01/01/1900'.to_date)
-          end
-          it 'sets default value for end_age_date' do
-            dob = 2.years.ago.to_date
-            as_dose_vaccine.end_age = nil
-            vaccine_attrs = test_target_dose.create_vaccine_attributes(
-              as_dose_vaccine,
-              dob
-            )
-            expect(vaccine_attrs[:end_age_date])
-              .to eq('12/31/2999'.to_date)
-          end
-        end
-      end
-
-      describe '#evaluate_vaccine_attributes' do
-        let(:valid_vaccine_attrs) do
-          {
-            begin_age_date: 10.months.ago.to_date,
-            end_age_date: 2.months.ago.to_date,
-            expected_trade_name: 'test',
-            expected_volume: '0.5'
-          }
-        end
-        describe 'for each minimum age attribute' do
-          administered_dose_date = 6.weeks.ago.to_date
-          attribute_options = {
-            before_the_dose_date: [8.weeks.ago.to_date, true],
-            after_the_dose_date: [4.weeks.ago.to_date, false],
-            nil: [nil, nil]
-          }
-          %w(
-            begin_age_date
-          ).each do |attribute|
-            attribute_options.each do |descriptor, value|
-              descriptor_string = "returns #{value[1]} when the #{attribute}"\
-                                  " attribute is #{descriptor}"
-              it descriptor_string do
-                valid_vaccine_attrs[attribute.to_sym] = value[0]
-                eval_hash = test_target_dose.evaluate_vaccine_attributes(
-                  valid_vaccine_attrs,
-                  administered_dose_date,
-                  'test',
-                  '0.5'
-                )
-                expect(eval_hash[attribute.to_sym]).to eq(value[1])
-              end
-            end
-          end
-        end
-        describe 'for each max age attribute' do
-          administered_dose_date = 6.weeks.ago.to_date
-          attribute_options = {
-            before_the_dose_date: [8.weeks.ago.to_date, false],
-            after_the_dose_date: [4.weeks.ago.to_date, true],
-            nil: [nil, nil]
-          }
-          %w(
-            end_age_date
-          ).each do |attribute|
-            attribute_options.each do |descriptor, value|
-              descriptor_string = "returns #{value[1]} when the #{attribute}"\
-                                  " attribute is #{descriptor}"
-              it descriptor_string do
-                valid_vaccine_attrs[attribute.to_sym] = value[0]
-                eval_hash = test_target_dose.evaluate_vaccine_attributes(
-                  valid_vaccine_attrs,
-                  administered_dose_date,
-                  'test',
-                  '0.5'
-                )
-                expect(eval_hash[attribute.to_sym]).to eq(value[1])
-              end
-            end
-          end
-        end
-        describe 'for the trade_name attribute' do
-          it 'returns true when dose trade_name equals expected_trade_name' do
-            valid_vaccine_attrs[:expected_trade_name] = 'tester'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'tester',
-              '0.5'
-            )
-            expect(eval_hash[:trade_name]).to eq(true)
-          end
-          it 'returns false when dose trade_name equals expected_trade_name' do
-            valid_vaccine_attrs[:expected_trade_name] = 'tester'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'NOT TESTER',
-              '0.5'
-            )
-            expect(eval_hash[:trade_name]).to eq(false)
-          end
-        end
-        describe 'for the volume attribute' do
-          it 'returns true when dose volume is equal to expected_volume' do
-            valid_vaccine_attrs[:expected_volume] = '0.5'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'tester',
-              '0.5'
-            )
-            expect(eval_hash[:volume]).to eq(true)
-          end
-          it 'returns true when dose volume is greater than expected_volume' do
-            valid_vaccine_attrs[:expected_volume] = '0.5'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'tester',
-              '0.9'
-            )
-            expect(eval_hash[:volume]).to eq(true)
-          end
-          it 'returns false when dose volume is less than expected_volume' do
-            valid_vaccine_attrs[:expected_volume] = '0.5'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'tester',
-              '0.3'
-            )
-            expect(eval_hash[:volume]).to eq(false)
-          end
-          it 'returns false when dose volume is nil' do
-            valid_vaccine_attrs[:expected_volume] = '0.5'
-            eval_hash = test_target_dose.evaluate_vaccine_attributes(
-              valid_vaccine_attrs,
-              Date.today,
-              'tester',
-              nil
-            )
-            expect(eval_hash[:volume]).to eq(false)
-          end
-        end
-      end
-
-      describe '#get_preferable_vaccine_status' do
-        # This logic is defined on page 50 of the CDC logic spec
-        it 'returns invalid, preferable, not_preferable for preferable false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'preferable',
-                              details: 'not_preferable' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns invalid, preferable, out_of_age_range for '\
-        'begin_age false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: false,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'preferable',
-                              details: 'out_of_age_range' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns invalid, preferable, out_of_age_range for '\
-        'end_age false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: false,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'preferable',
-                              details: 'out_of_age_range' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns invalid, preferable, wrong_trade_name for '\
-        'trade_name false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: false,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'preferable',
-                              details: 'wrong_trade_name' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns valid, preferable, less_than_recommended_volume for ' \
-          'volume false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: true,
-            volume: false
-          }
-          expected_result = { status: 'valid',
-                              reason: 'preferable',
-                              details: 'less_than_recommended_volume' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-        it 'returns valid, preferable for all true' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'valid',
-                              reason: 'preferable' }
-          expect(
-            test_target_dose.get_preferable_vaccine_status(vaccine_eval_hash,
-                                                           prev_status_hash)
-          ).to eq(expected_result)
-        end
-      end
-
-      describe '#get_allowable_vaccine_status' do
-        # This logic is defined on page 52 of the CDC logic spec
-        it 'returns invalid, allowable, not_allowable for allowable false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'preferable',
-                              details: 'not_preferable' }
-          expect(
-            test_target_dose.get_allowable_vaccine_status(vaccine_eval_hash,
-                                                          prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns invalid, allowable, out_of_age_range for '\
-        'begin_age false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: false,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'allowable',
-                              details: 'out_of_age_range' }
-          expect(
-            test_target_dose.get_allowable_vaccine_status(vaccine_eval_hash,
-                                                          prev_status_hash)
-          ).to eq(expected_result)
-        end
-
-        it 'returns invalid, allowable, out_of_age_range for '\
-        'end_age false' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: false,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'invalid',
-                              reason: 'allowable',
-                              details: 'out_of_age_range' }
-          expect(
-            test_target_dose.get_allowable_vaccine_status(vaccine_eval_hash,
-                                                          prev_status_hash)
-          ).to eq(expected_result)
-        end
-        it 'returns valid, allowable for all true' do
-          prev_status_hash = nil
-          vaccine_eval_hash = {
-            begin_age: true,
-            end_age: true,
-            trade_name: true,
-            volume: true
-          }
-          expected_result = { status: 'valid',
-                              reason: 'allowable' }
-          expect(
-            test_target_dose.get_allowable_vaccine_status(vaccine_eval_hash,
-                                                          prev_status_hash)
-          ).to eq(expected_result)
-        end
       end
     end
     describe 'implementation of evaluate gender logic' do
